Day 18, part 2

+++ 2023/Day18.h
@@ -7,7 +7,7 @@ private:
 	const std::string testFile2 = "inputs/day18_part2_test.txt";
 	const std::string puzzleFile = "inputs/day18.txt";
 	int calculatePuzzle1(std::vector<std::string> input);
-	int calculatePuzzle2(std::vector<std::string> input);
+	uint64_t calculatePuzzle2(std::vector<std::string> input);
 public:
 	void puzzle1();
 	void puzzle2();
